--- cypress/support/index.js
@@ -1,2 +1,0 @@
-import 'cypress-hmr-restarter'
-import '@testing-library/cypress/add-commands'
