+++ src/setupTests.extra-4.js
@@ -1,0 +1,24 @@
+import '@testing-library/jest-dom/extend-expect'
+import {queryCache} from 'react-query'
+import * as auth from 'auth-provider'
+import {server} from 'test/server'
+import * as usersDB from 'test/data/users'
+import * as listItemsDB from 'test/data/list-items'
+import * as booksDB from 'test/data/books'
+
+// enable API mocking in test runs using the same request handlers
+// as for the client-side mocking.
+beforeAll(() => server.listen())
+afterAll(() => server.close())
+afterEach(() => server.resetHandlers())
+
+// general cleanup
+afterEach(async () => {
+  queryCache.clear()
+  await Promise.all([
+    auth.logout(),
+    usersDB.reset(),
+    booksDB.reset(),
+    listItemsDB.reset(),
+  ])
+})
