+++ src/setupTests.extra-5.js
@@ -1,0 +1,39 @@
+import '@testing-library/jest-dom/extend-expect'
+import {act} from '@testing-library/react'
+import {queryCache} from 'react-query'
+import * as auth from 'auth-provider'
+import {server} from 'test/server'
+import * as usersDB from 'test/data/users'
+import * as listItemsDB from 'test/data/list-items'
+import * as booksDB from 'test/data/books'
+
+// we don't need the profiler in tests
+jest.mock('components/profiler')
+
+// enable API mocking in test runs using the same request handlers
+// as for the client-side mocking.
+beforeAll(() => server.listen())
+afterAll(() => server.close())
+afterEach(() => server.resetHandlers())
+
+// general cleanup
+afterEach(async () => {
+  queryCache.clear()
+  await Promise.all([
+    auth.logout(),
+    usersDB.reset(),
+    booksDB.reset(),
+    listItemsDB.reset(),
+  ])
+})
+
+// real times is a good default to start, individual tests can
+// enable fake timers if they need.
+// it's important this comes last here because jest runs afterEach callbacks
+// in reverse order and we want this to be run first.
+afterEach(() => {
+  if (jest.isMockFunction(setTimeout)) {
+    act(() => jest.runOnlyPendingTimers())
+    jest.useRealTimers()
+  }
+})
